Integrate configtree components.
* Align configtree parsers and combiners with similar ones that already exist.
* Extend ini, yaml, and json base classes to support configtree style navigation.
