+++ docs/notebook_index.rst
@@ -8,6 +8,7 @@ directory.
 .. toctree::
     :titlesonly:
 
+    notebooks/Config Tree Tutorial.ipynb
     notebooks/Datasource Registry.ipynb
     notebooks/Development Examples.ipynb
     notebooks/Diagnostic Walkthrough.ipynb
