+++ insights/parsers/tests/test_multipath_conf.py
@@ -1,5 +1,19 @@
+from insights.configtree import first, last
 from insights.parsers import multipath_conf
 from insights.tests import context_wrap
+
+
+CONF = """
+blacklist {
+       device {
+               vendor  "IBM"
+               product "3S42"       #DS4200 Product 10
+       }
+       device {
+               vendor  "HP"
+               product "*"
+       }
+}""".strip()
 
 
 MULTIPATH_CONF_INFO = """
@@ -46,3 +60,18 @@ def test_multipath_conf():
     assert multipath_conf_info.get('multipaths')[1].get('alias') == 'red'
     assert multipath_conf_info.get('devices')[0].get('no_path_retry') == 'queue'
     assert multipath_conf_info.get('blacklist').get('devnode') == '^hd[a-z]'
+
+
+def test_multipath_conf_tree():
+    conf = multipath_conf.MultipathConfTree(context_wrap(CONF))
+    assert len(conf["blacklist"]) == 1
+    assert len(conf["blacklist"]["device"]) == 2
+
+    assert len(conf["blacklist"]["device"]["vendor"]) == 2
+    assert len(conf["blacklist"]["device"]["product"]) == 2
+
+    assert conf["blacklist"]["device"]["vendor"][first].value == "IBM"
+    assert conf["blacklist"]["device"]["vendor"][last].value == "HP"
+
+    assert conf["blacklist"]["device"]["product"][first].value == "3S42"
+    assert conf["blacklist"]["device"]["product"][last].value == "*"
